Interim commit with the first stage of remote connectivity.
Part of the work for #1

This simply makes url requests to a public website and appends the
source url as an unencoded query parameter.

Tested interactively, the log messages indicate the outbound http
requests work and that we receive a web page in response.

This code will be updated shortly to make requests to another
running emulator instance.