The next step in inter-emulator communications, they communicate!
Part of #1.

So this is a bit crude, nonetheless emulators can now communicate provided
the string 'askremoteserver' is somewhere in the URL that's received.

Adding this has exposed a key limitation in the current codebase which
checks the end of the URL to try and determine what to serve. The current
code doesn't process the URI e.g. for query string parameters. I've created #5
to track this work.

Thanks to https://stackoverflow.com/questions/67538590/cleartext-http-traffic-to-not-permitted
when enabled me to disable the plaintext blocking Android performs by default.
It saves me from trying to add SSL support to this codebase.

The functionality has been tested interactively using wget and in firefox.
The adb logs provide additional evidence that the two instances are communicating.